# docs/api/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "WhatsApp AI Agent Platform API",
    "description": "Production-grade SaaS WhatsApp AI Agent Platform with Multi-Agent & Graph RAG architecture",
    "version": "1.0.0",
    "contact": {
      "name": "API Support",
      "email": "api-support@whatsappagent.io"
    },
    "license": {
      "name": "Proprietary"
    }
  },
  "servers": [
    {
      "url": "/api",
      "description": "API Gateway"
    }
  ],
...
  "paths": {
...
    "/reminders": {
      "get": {
        "tags": ["Reminders"],
        "summary": "List reminders",
        "description": "Get all reminders for the authenticated user",
        "operationId": "listReminders",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "description": "Filter by status",
            "schema": {
              "type": "string",
              "enum": ["pending", "completed", "cancelled"]
            }
          },
          {
            "name": "from",
            "in": "query",
            "description": "Start date filter",
            "schema": { "type": "string", "format": "date-time" }
          },
          {
            "name": "to",
            "in": "query",
            "description": "End date filter",
            "schema": { "type": "string", "format": "date-time" }
          }
        ],
        "responses": {
          "200": {
            "description": "Reminders retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "reminders": {
                          "type": "array",
                          "items": { "$ref": "#/components/schemas/Reminder" }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Reminders"],
        "summary": "Create reminder",
        "description": "Create a new reminder",
        "operationId": "createReminder",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReminderRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Reminder created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "reminder": { "$ref": "#/components/schemas/Reminder" }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/reminders/{id}": {
      "get": {
        "tags": ["Reminders"],
        "summary": "Get reminder",
        "description": "Get a specific reminder by ID",
        "operationId": "getReminder",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "Reminder ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Reminder retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "reminder": { "$ref": "#/components/schemas/Reminder" }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Reminder not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Reminders"],
        "summary": "Update reminder",
        "description": "Update an existing reminder",
        "operationId": "updateReminder",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "Reminder ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateReminderRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Reminder updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "reminder": { "$ref": "#/components/schemas/Reminder" }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Reminder not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Reminders"],
        "summary": "Delete reminder",
        "description": "Delete a reminder",
        "operationId": "deleteReminder",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "Reminder ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Reminder deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "404": {
            "description": "Reminder not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/reminders/{id}/complete": {
      "post": {
        "tags": ["Reminders"],
        "summary": "Complete reminder",
        "description": "Mark a reminder as completed",
        "operationId": "completeReminder",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "Reminder ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Reminder completed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "reminder": { "$ref": "#/components/schemas/Reminder" }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Reminder not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/reminders/{id}/share": {
      "post": {
        "tags": ["Reminders"],
        "summary": "Share reminder",
        "description": "Share a reminder with another user",
        "operationId": "shareReminder",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "Reminder ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShareReminderRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Reminder shared",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "sharedReminder": { "$ref": "#/components/schemas/SharedReminder" }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/reminders/shared/with-me": {
      "get": {
        "tags": ["Reminders"],
        "summary": "Get shared reminders",
        "description": "Get reminders shared with the authenticated user",
        "operationId": "getSharedReminders",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Shared reminders retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "reminders": {
                          "type": "array",
                          "items": {
                            "allOf": [
                              { "$ref": "#/components/schemas/Reminder" },
                              {
                                "type": "object",
                                "properties": {
                                  "permission": { "type": "string", "enum": ["view", "edit"] },
                                  "sharedAt": { "type": "string", "format": "date-time" }
                                }
                              }
                            ]
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT authentication token"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "object",
            "properties": {
              "code": { "type": "string" },
              "message": { "type": "string" }
            }
          }
        }
      },
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object",
            "properties": {
              "message": { "type": "string" }
            }
          }
        }
      },
...
      "Reminder": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "userId": { "type": "string" },
          "title": { "type": "string" },
          "description": { "type": "string" },
          "scheduledAt": { "type": "string", "format": "date-time" },
          "timezone": { "type": "string" },
          "repeatType": { "type": "string", "enum": ["NONE", "DAILY", "WEEKLY", "MONTHLY", "YEARLY"] },
          "repeatInterval": { "type": "integer" },
          "repeatEndDate": { "type": "string", "format": "date-time" },
          "status": { "type": "string", "enum": ["PENDING", "COMPLETED", "CANCELLED"] },
          "metadata": { "type": "object" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "CreateReminderRequest": {
        "type": "object",
        "required": ["title", "scheduledAt"],
        "properties": {
          "title": { "type": "string", "minLength": 1 },
          "description": { "type": "string" },
          "scheduledAt": { "type": "string", "format": "date-time" },
          "timezone": { "type": "string", "default": "UTC" },
          "repeatType": { "type": "string", "enum": ["none", "daily", "weekly", "monthly", "yearly"], "default": "none" },
          "repeatInterval": { "type": "integer", "minimum": 1, "default": 1 },
          "repeatEndDate": { "type": "string", "format": "date-time" },
          "metadata": { "type": "object" }
        }
      },
      "UpdateReminderRequest": {
        "type": "object",
        "properties": {
          "title": { "type": "string", "minLength": 1 },
          "description": { "type": "string" },
          "scheduledAt": { "type": "string", "format": "date-time" },
          "timezone": { "type": "string" },
          "repeatType": { "type": "string", "enum": ["none", "daily", "weekly", "monthly", "yearly"] },
          "repeatInterval": { "type": "integer", "minimum": 1 },
          "repeatEndDate": { "type": "string", "format": "date-time" },
          "status": { "type": "string", "enum": ["pending", "completed", "cancelled"] },
          "metadata": { "type": "object" }
        }
      },
      "ShareReminderRequest": {
        "type": "object",
        "required": ["userId"],
        "properties": {
          "userId": { "type": "string", "format": "uuid", "description": "User ID to share with" },
          "permission": { "type": "string", "enum": ["view", "edit"], "default": "view" }
        }
      },
      "SharedReminder": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "reminderId": { "type": "string" },
          "sharedWithUserId": { "type": "string" },
          "permission": { "type": "string", "enum": ["view", "edit"] },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
...
    }
  }
}